Move working code out into src/
